pyproject.toml
[project]
name = "dead-leaves"
version = "0.1.0"
description = "Simple dead leaves synthesis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glfw>=2.9.0",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "pyglm>=2.8.2",
    "pyopengl>=3.1.9",
]
